Separate ServerResourceFormatter into server-specific formatters
Ebenezer no longer needs the Win32 formatter and AI never needed the DB formatter.
They can just implement their own, there's no real reason to share this anymore.
