+++ Server/AIServer/AIServer.vcxproj
@@ -164,6 +164,7 @@
     <ClInclude Include="..\..\Client\N3Base\My_3DStruct.h" />
     <ClInclude Include="..\..\Client\N3Base\N3ShapeMgr.h" />
     <ClInclude Include="..\..\Client\N3Base\StdAfxBase.h" />
+    <ClInclude Include="AIResourceFormatter.h" />
     <ClInclude Include="Define.h" />
     <ClInclude Include="Extern.h" />
     <ClInclude Include="GameSocket.h" />
