+++ Server/AIServer/AIServer.vcxproj.filters
@@ -157,6 +157,9 @@
     <ClInclude Include="..\..\Client\N3Base\StdAfxBase.h">
       <Filter>N3Base</Filter>
     </ClInclude>
+    <ClInclude Include="AIResourceFormatter.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <Image Include="res\Server.ico">
