+++ Server/Ebenezer/Ebenezer.vcxproj.filters
@@ -81,7 +81,7 @@
     <ClCompile Include="..\..\Client\N3Base\N3ShapeMgr.cpp">
       <Filter>N3Base</Filter>
     </ClCompile>
-    <ClCompile Include="ServerResourceFormatter.cpp">
+    <ClCompile Include="EbenezerResourceFormatter.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
@@ -168,6 +168,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="db_resources.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="EbenezerResourceFormatter.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
